chore: move getItemIndex to AbstractDataView
Moved `getItemIndex(Object)` and `getItemIndex(Object, Stream)` methods from `AbstractListDataView` to `AbstractDataView`.
